fix(camera)!: move to grinder/ subpkg
