+++ lib/src/main/java/io/github/nonmilk/coffee/grinder/camera/OrthographicCamera.java
@@ -1,4 +1,4 @@
-package io.github.nonmilk.coffee.camera;
+package io.github.nonmilk.coffee.grinder.camera;
 
 import io.github.alphameo.linear_algebra.mat.Matrix4;
